Fix not to encode response when the status indicates the content is empty Motivation: `EncodingService` should not try to encode the body when the status indicates that the content it empty.
Modifications:
- Check the status if it indicates the content is empty in `EncodingService`.
- Downgrade Micrometer to 1.3.6 becuase 1.4.0 is Non-LTS.

Result:
- `EncodingService` does not encode the content when the status is 204, 205 and 304

+++ core/src/main/java/com/linecorp/armeria/common/metric/MoreMeters.java
@@ -112,12 +112,12 @@ public final class MoreMeters {
         requireNonNull(name, "name");
         requireNonNull(tags, "tags");
 
-        final Double maxExpectedValueNanos = distStatCfg.getMaximumExpectedValue();
-        final Double minExpectedValueNanos = distStatCfg.getMinimumExpectedValue();
+        final Long maxExpectedValueNanos = distStatCfg.getMaximumExpectedValue();
+        final Long minExpectedValueNanos = distStatCfg.getMinimumExpectedValue();
         final Duration maxExpectedValue =
-                maxExpectedValueNanos != null ? Duration.ofNanos(maxExpectedValueNanos.longValue()) : null;
+                maxExpectedValueNanos != null ? Duration.ofNanos(maxExpectedValueNanos) : null;
         final Duration minExpectedValue =
-                minExpectedValueNanos != null ? Duration.ofNanos(minExpectedValueNanos.longValue()) : null;
+                minExpectedValueNanos != null ? Duration.ofNanos(minExpectedValueNanos) : null;
 
         return Timer.builder(name)
                     .tags(tags)
